Moved packs and encounter sets out to the api Added ability of the plugin to provide a list of packs to the program
