+++ EideticMemoryOverlay.PluginApi/Pack.cs
@@ -1,0 +1,32 @@
+﻿using System.Collections.Generic;
+
+namespace EideticMemoryOverlay.PluginApi {
+    public class Pack {
+        public Pack() {
+            EncounterSets = new List<EncounterSet>();
+        }
+
+        public Pack(Pack pack) {
+            Code = pack.Code;
+            Name = pack.Name;
+            CyclePosition = pack.CyclePosition;
+            Position = pack.Position;
+
+            EncounterSets = new List<EncounterSet>();
+            foreach (var encounterSet in pack.EncounterSets) {
+                EncounterSets.Add(new EncounterSet(encounterSet));
+            }
+        }
+
+        public string Code { get; set; }
+
+        public string Name { get; set; }
+
+        public int CyclePosition { get; set; }
+
+        public int Position { get; set; }
+
+        public IList<EncounterSet> EncounterSets { get; set; }
+    }
+
+}
