+++ Plugins/ArkhamHorrorLcg/ArkhamConfiguration.cs
@@ -61,33 +61,4 @@ namespace ArkhamHorrorLcg {
         public IList<Pack> Packs { get; set; }
         public string LocalImagesDirectory { get; set; }
     }
-
-    public class Pack {
-        public Pack() {
-            EncounterSets = new List<EncounterSet>();
-        }
-
-        public Pack(Pack pack) {
-            Code = pack.Code;
-            Name = pack.Name;
-            CyclePosition = pack.CyclePosition;
-            Position = pack.Position;
-
-            EncounterSets = new List<EncounterSet>();
-            foreach (var encounterSet in pack.EncounterSets) {
-                EncounterSets.Add(new EncounterSet(encounterSet));
-            }
-        }
-
-        public string Code { get; set; }
-
-        public string Name { get; set; }
-
-        public int CyclePosition { get; set; }
-
-        public int Position { get; set; }
-
-        public IList<EncounterSet> EncounterSets { get; set; }
-    }
-
 }
